+++ hypr/components/input.lua
@@ -1,0 +1,25 @@
+-------------
+--- INPUT ---
+-------------
+
+-- https://wiki.hypr.land/Configuring/Basics/Variables/
+hl.config({
+    input = {
+        kb_layout  = "us",
+        kb_variant = "",
+        kb_model   = "",
+        kb_options = "caps:super",
+        kb_rules   = "",
+
+        follow_mouse = 1,
+
+        sensitivity = 0, -- -1.0 - 1.0, 0 means no modification.
+    },
+})
+
+-- See https://wiki.hypr.land/Configuring/Gestures
+hl.gesture({
+    fingers   = 3,
+    direction = "horizontal",
+    action    = "workspace",
+})